imbulichka/requirements.txt
discord.py==2.3.2
aiosqlite==0.19.0
python-dotenv==1.0.0
openai==1.3.0
aiohttp==3.9.0
yt-dlp==2023.10.13
Pillow==10.1.0
numpy==1.24.3